# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Omni - Documentação da API",
        "description": "Documentação de todos os endpoints da API do sistema Omni.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/admin/files-omnis": {
            "get": {
                "tags": [
                    "Files Omni"
                ],
                "summary": "Lista todos os arquivos",
                "operationId": "a0ed84a3db126fda85b61b05616c5b3b",
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/FilesOmni"
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Files Omni"
                ],
                "summary": "Cria um novo registro de arquivo",
                "operationId": "3b939cbae11e24467d17d9958e15cd8d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreFilesOmniRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Criado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesOmni"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/files-omnis/{id}": {
            "get": {
                "tags": [
                    "Files Omni"
                ],
                "summary": "Busca um arquivo pelo ID",
                "operationId": "a0d7efd1cc09b8667b518ce3d95c6b2e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesOmni"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Files Omni"
                ],
                "summary": "Atualiza um registro de arquivo",
                "operationId": "a545de7d1b23661f840898237140ceef",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateFilesOmniRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FilesOmni"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Arquivo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Files Omni"
                ],
                "summary": "Deleta um registro de arquivo",
                "operationId": "579ae5286c389cd9375d246424344718",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Deletado com sucesso"
                    },
                    "404": {
                        "description": "Arquivo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "StoreFilesOmniRequest": {
                "title": "Store FilesOmni Request",
                "required": [
                    "task_id",
                    "client_id",
                    "uploaded_by",
                    "file_path",
                    "file_name",
                    "file_type",
                    "version"
                ],
                "properties": {
                    "task_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "client_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "uploaded_by": {
                        "type": "integer",
                        "example": 1
                    },
                    "file_path": {
                        "type": "string",
                        "example": "uploads/clients/file123.pdf"
                    },
                    "file_name": {
                        "type": "string",
                        "example": "contrato.pdf"
                    },
                    "file_type": {
                        "type": "string",
                        "example": "application/pdf"
                    },
                    "version": {
                        "type": "string",
                        "enum": [
                            "preview",
                            "final",
                            "for_color"
                        ],
                        "example": "final"
                    },
                    "video_orientation": {
                        "type": "string",
                        "enum": [
                            "vertical",
                            "horizontal"
                        ],
                        "example": "vertical"
                    },
                    "cover_image": {
                        "type": "string",
                        "example": "path/to/cover.jpg"
                    },
                    "published": {
                        "type": "boolean",
                        "example": true
                    }
                },
                "type": "object"
            },
            "UpdateFilesOmniRequest": {
                "properties": {
                    "file_name": {
                        "type": "string",
                        "example": "contrato_assinado.pdf"
                    },
                    "published": {
                        "type": "boolean",
                        "example": false
                    }
                },
                "type": "object"
            },
...
            "FilesOmni": {
                "title": "Files Omni",
                "description": "Representa um arquivo genérico no sistema, associado a um cliente e opcionalmente a uma tarefa.",
                "properties": {
                    "id": {
                        "description": "ID único do arquivo",
                        "type": "integer",
                        "readOnly": true
                    },
                    "task_id": {
                        "description": "ID da tarefa associada (opcional)",
                        "type": "integer",
                        "nullable": true
                    },
                    "client_id": {
                        "description": "ID do cliente associado",
                        "type": "integer"
                    },
                    "uploaded_by": {
                        "description": "ID do usuário que fez o upload",
                        "type": "integer"
                    },
                    "file_path": {
                        "description": "Caminho do arquivo no storage",
                        "type": "string",
                        "example": "clients/1/tasks/123/video.mp4"
                    },
                    "file_name": {
                        "description": "Nome original do arquivo",
                        "type": "string",
                        "example": "video_casamento.mp4"
                    },
                    "file_type": {
                        "description": "MIME type do arquivo",
                        "type": "string",
                        "example": "video/mp4"
                    },
                    "version": {
                        "description": "Versão do arquivo (ex: original, editado)",
                        "type": "string",
                        "enum": [
                            "original",
                            "editado",
                            "preview"
                        ],
                        "example": "original"
                    },
                    "cover_image": {
                        "description": "Caminho para a imagem de capa do arquivo (geralmente para vídeos)",
                        "type": "string",
                        "example": "clients/1/tasks/123/cover.jpg",
                        "nullable": true
                    },
                    "video_orientation": {
                        "description": "Orientação do vídeo",
                        "type": "string",
                        "enum": [
                            "landscape",
                            "portrait"
                        ],
                        "example": "landscape",
                        "nullable": true
                    },
                    "published": {
                        "description": "Indica se o arquivo está publicado e visível para o cliente",
                        "type": "boolean",
                        "example": true
                    },
                    "created_at": {
                        "description": "Data de criação",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    },
                    "updated_at": {
                        "description": "Data da última atualização",
                        "type": "string",
                        "format": "date-time",
                        "readOnly": true
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}